+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.2 — Unreleased
 
-- 
+- Renamed the `--address` option to the command-line client to `--server-url`
 
 ### 0.1 — July 22, 2016